# packages/backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Patrinotes API","description":"Patrinotes API Documentation","version":"1.0"},"components":{"schemas":{}},"paths":{"/health":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/authenticate":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"code":{"type":"string"}},"required":["code"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"tokens":{"type":"object","properties":{"access_token":{"type":"string"},"expires_in":{"type":"string"},"refresh_token":{"type":"string"},"refresh_expires_in":{"type":"string"},"token_type":{"type":"string"},"session_state":{"type":"string"},"scope":{"type":"string"},"id_token":{"type":"string"}},"required":["access_token","expires_in","refresh_token","refresh_expires_in","token_type","session_state","scope","id_token"]},"user":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"service_id":{"type":"string"},"service":{"type":"object","properties":{"id":{"type":"string"},"department":{"type":"string"},"completeCoords":{"anyOf":[{"type":"string"},{"type":"null"}]},"visible":{"anyOf":[{"type":"boolean"},{"type":"null"}]},"name":{"anyOf":[{"type":"string"},{"type":"null"}]},"address":{"anyOf":[{"type":"string"},{"type":"null"}]},"zipCode":{"anyOf":[{"type":"string"},{"type":"null"}]},"city":{"anyOf":[{"type":"string"},{"type":"null"}]},"phone":{"anyOf":[{"type":"string"},{"type":"null"}]},"email":{"anyOf":[{"type":"string"},{"type":"null"}]},"marianne_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"drac_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"dept_numbers":{"anyOf":[{"type":"string"},{"type":"null"}]},"service_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_crmh":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_caoa":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_dreal":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_sra":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_udap":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_ofb":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","department"]},"job":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","name","service_id","service","job"]}},"required":["tokens","user"]}}}}}}},"/api/refresh-token":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"user":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"service_id":{"type":"string"},"service":{"type":"object","properties":{"id":{"type":"string"},"department":{"type":"string"},"completeCoords":{"anyOf":[{"type":"string"},{"type":"null"}]},"visible":{"anyOf":[{"type":"boolean"},{"type":"null"}]},"name":{"anyOf":[{"type":"string"},{"type":"null"}]},"address":{"anyOf":[{"type":"string"},{"type":"null"}]},"zipCode":{"anyOf":[{"type":"string"},{"type":"null"}]},"city":{"anyOf":[{"type":"string"},{"type":"null"}]},"phone":{"anyOf":[{"type":"string"},{"type":"null"}]},"email":{"anyOf":[{"type":"string"},{"type":"null"}]},"marianne_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"drac_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"dept_numbers":{"anyOf":[{"type":"string"},{"type":"null"}]},"service_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_crmh":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_caoa":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_dreal":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_sra":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_udap":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_ofb":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","department"]},"job":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","name","service_id","service","job"]},"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"expiresAt":{"type":"string"}},"required":["user","accessToken","refreshToken","expiresAt"]}}}}}}},"/api/create-user":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"password":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"job":{"type":"string"},"service_id":{"type":"string"},"newsletter":{"type":"boolean"},"cgu":{"type":"boolean"}},"required":["password","name","email","job","service_id","newsletter","cgu"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"user":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"service_id":{"type":"string"},"service":{"type":"object","properties":{"id":{"type":"string"},"department":{"type":"string"},"completeCoords":{"anyOf":[{"type":"string"},{"type":"null"}]},"visible":{"anyOf":[{"type":"boolean"},{"type":"null"}]},"name":{"anyOf":[{"type":"string"},{"type":"null"}]},"address":{"anyOf":[{"type":"string"},{"type":"null"}]},"zipCode":{"anyOf":[{"type":"string"},{"type":"null"}]},"city":{"anyOf":[{"type":"string"},{"type":"null"}]},"phone":{"anyOf":[{"type":"string"},{"type":"null"}]},"email":{"anyOf":[{"type":"string"},{"type":"null"}]},"marianne_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"drac_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"dept_numbers":{"anyOf":[{"type":"string"},{"type":"null"}]},"service_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_crmh":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_caoa":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_dreal":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_sra":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_udap":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_ofb":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","department"]},"job":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","name","service_id","service","job"]},"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"expiresAt":{"type":"string"}},"required":["user","accessToken","refreshToken","expiresAt"]}}}}}}},"/api/login-user":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"user":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"service_id":{"type":"string"},"service":{"type":"object","properties":{"id":{"type":"string"},"department":{"type":"string"},"completeCoords":{"anyOf":[{"type":"string"},{"type":"null"}]},"visible":{"anyOf":[{"type":"boolean"},{"type":"null"}]},"name":{"anyOf":[{"type":"string"},{"type":"null"}]},"address":{"anyOf":[{"type":"string"},{"type":"null"}]},"zipCode":{"anyOf":[{"type":"string"},{"type":"null"}]},"city":{"anyOf":[{"type":"string"},{"type":"null"}]},"phone":{"anyOf":[{"type":"string"},{"type":"null"}]},"email":{"anyOf":[{"type":"string"},{"type":"null"}]},"marianne_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"drac_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"dept_numbers":{"anyOf":[{"type":"string"},{"type":"null"}]},"service_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_crmh":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_caoa":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_dreal":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_sra":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_udap":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_ofb":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","department"]},"job":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","name","service_id","service","job"]},"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"expiresAt":{"type":"string"}},"required":["user","accessToken","refreshToken","expiresAt"]}}}}}}},"/api/change-service":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"service_id":{"type":"string"}},"required":["service_id"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]}}}}}}},"/api/created-at":{"get":{"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"createdAt":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["createdAt"]}}}}}}},"/api/send-reset-password":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]}}}}}}},"/api/reset-password":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"temporaryLink":{"type":"string"},"newPassword":{"type":"string"}},"required":["temporaryLink","newPassword"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]}}}}}}},"/api/services":{"get":{"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"department":{"type":"string"},"completeCoords":{"anyOf":[{"type":"string"},{"type":"null"}]},"visible":{"anyOf":[{"type":"boolean"},{"type":"null"}]},"name":{"anyOf":[{"type":"string"},{"type":"null"}]},"address":{"anyOf":[{"type":"string"},{"type":"null"}]},"zipCode":{"anyOf":[{"type":"string"},{"type":"null"}]},"city":{"anyOf":[{"type":"string"},{"type":"null"}]},"phone":{"anyOf":[{"type":"string"},{"type":"null"}]},"email":{"anyOf":[{"type":"string"},{"type":"null"}]},"marianne_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"drac_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"dept_numbers":{"anyOf":[{"type":"string"},{"type":"null"}]},"service_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_crmh":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_caoa":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_dreal":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_sra":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_udap":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_ofb":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","department"]}}}}}}}},"/api/upload/attachment/presigned-url":{"get":{"parameters":[{"schema":{"type":"string"},"in":"query","name":"filePath","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"url":{"type":"string"}},"required":["url"]}}}}}}},"/api/upload/attachment":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/upload/attachment/size":{"get":{"responses":{"200":{"description":"Default Response"}}}},"/api/pdf/report/upload-url":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"reportId":{"type":"string"}},"required":["reportId"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"uploadUrl":{"type":"string"},"pdfPath":{"type":"string"}},"required":["uploadUrl","pdfPath"]}}}}}}},"/api/pdf/report":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"pdfPath":{"type":"string"},"pdfSize":{"type":"number"},"reportId":{"type":"string"},"recipients":{"type":"string"}},"required":["pdfPath","reportId","recipients"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"string"}}}}}},"get":{"parameters":[{"schema":{"type":"string"},"in":"query","name":"reportId","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}}}},"/api/pdf/state-report":{"get":{"parameters":[{"schema":{"type":"string"},"in":"query","name":"stateReportId","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}}},"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"needValidation":{"type":"boolean"},"pdfPath":{"type":"string"},"pdfSize":{"type":"number"},"stateReportId":{"type":"string"},"recipients":{"type":"string"},"alerts":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"alert":{"anyOf":[{"type":"string"},{"type":"null"}]},"commentaires":{"anyOf":[{"type":"string"},{"type":"null"}]},"show_in_report":{"anyOf":[{"type":"boolean"},{"type":"number"},{"type":"null"}]},"mandatory_emails":{"anyOf":[{"type":"string"},{"type":"null"}]},"additional_emails":{"anyOf":[{"type":"string"},{"type":"null"}]},"objet_ou_mobilier":{"anyOf":[{"type":"string"},{"type":"null"}]},"objet_ou_mobilier_name":{"anyOf":[{"type":"string"},{"type":"null"}]},"probleme":{"anyOf":[{"type":"string"},{"type":"null"}]},"should_send":{"anyOf":[{"type":"boolean"},{"type":"number"},{"type":"null"}]}},"required":["id","alert","commentaires","show_in_report","mandatory_emails","additional_emails","objet_ou_mobilier","objet_ou_mobilier_name","probleme","should_send"]}}},"required":["pdfPath","stateReportId","recipients"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/pdf/state-report/upload-url":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"stateReportId":{"type":"string"}},"required":["stateReportId"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"uploadUrl":{"type":"string"},"pdfPath":{"type":"string"}},"required":["uploadUrl","pdfPath"]}}}}}}},"/api/state-report/objets-images":{"get":{"parameters":[{"schema":{"type":"string"},"in":"query","name":"references","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"reference":{"anyOf":[{"type":"string"},{"type":"null"}]},"url":{"anyOf":[{"type":"string"},{"type":"null"}]},"dept_number":{"anyOf":[{"type":"string"},{"type":"null"}]},"label":{"anyOf":[{"type":"string"},{"type":"null"}]},"copyright":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id"]}}}}}}}},"/api/constat-validation/{token}":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"token","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}}}},"/api/constat-validation/{token}/pdf":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"token","required":true}],"responses":{"200":{"description":"Default Response"}}}},"/api/constat-validation/{token}/accept":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"comment":{"type":"string"}}}}}},"parameters":[{"schema":{"type":"string"},"in":"path","name":"token","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]}}}}}}},"/api/constat-validation/{token}/decline":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"comment":{"type":"string"}},"required":["comment"]}}},"required":true},"parameters":[{"schema":{"type":"string"},"in":"path","name":"token","required":true}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]}}}}}}},"/api/upload-data":{"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"op_id":{"type":"number"},"tx_id":{"anyOf":[{"type":"number"},{"type":"null"}]},"id":{"type":"string"},"type":{"type":"string"},"op":{"type":"string"},"data":{}},"required":["op_id","id","type","op"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}}}},"/api/admin/me":{"get":{"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"service_id":{"type":"string"},"service":{"type":"object","properties":{"id":{"type":"string"},"department":{"type":"string"},"completeCoords":{"anyOf":[{"type":"string"},{"type":"null"}]},"visible":{"anyOf":[{"type":"boolean"},{"type":"null"}]},"name":{"anyOf":[{"type":"string"},{"type":"null"}]},"address":{"anyOf":[{"type":"string"},{"type":"null"}]},"zipCode":{"anyOf":[{"type":"string"},{"type":"null"}]},"city":{"anyOf":[{"type":"string"},{"type":"null"}]},"phone":{"anyOf":[{"type":"string"},{"type":"null"}]},"email":{"anyOf":[{"type":"string"},{"type":"null"}]},"marianne_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"drac_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"dept_numbers":{"anyOf":[{"type":"string"},{"type":"null"}]},"service_text":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_crmh":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_caoa":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_dreal":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_sra":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_udap":{"anyOf":[{"type":"string"},{"type":"null"}]},"courriel_ofb":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","department"]},"job":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["id","name","service_id","service","job"]}}}}}}},"/api/admin/whitelist":{"get":{"parameters":[{"schema":{"default":1,"type":"number"},"in":"query","name":"page","required":false},{"schema":{"default":20,"type":"number"},"in":"query","name":"limit","required":false}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"type":"object","properties":{"email":{"type":"string"},"createdAt":{"type":"string"},"hasUser":{"type":"boolean"},"lastCreatedStateReport":{"anyOf":[{"type":"string"},{"type":"null"}]},"lastFinishedStateReport":{"anyOf":[{"type":"string"},{"type":"null"}]}},"required":["email","createdAt","hasUser","lastCreatedStateReport","lastFinishedStateReport"]}},"total":{"type":"number"},"page":{"type":"number"},"limit":{"type":"number"}},"required":["data","total","page","limit"]}}}}}},"post":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]}}}}}},"delete":{"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]}}},"required":true},"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]}}}}}}},"/api/admin/whitelist/export":{"get":{"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"email":{"type":"string"},"createdAt":{"type":"string"}},"required":["email","createdAt"]}}}}}}}},"/api/admin/users/export":{"get":{"parameters":[{"schema":{"type":"string"},"in":"query","name":"search","required":false}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"job":{"anyOf":[{"type":"string"},{"type":"null"}]},"serviceId":{"type":"string"},"serviceName":{"anyOf":[{"type":"string"},{"type":"null"}]},"serviceDepartment":{"anyOf":[{"type":"string"},{"type":"null"}]},"role":{"anyOf":[{"type":"string"},{"type":"null"}]},"createdAt":{"type":"string"}},"required":["id","name","email","job","serviceId","serviceName","serviceDepartment","role","createdAt"]}}}}}}}},"/api/admin/users":{"get":{"parameters":[{"schema":{"default":1,"type":"number"},"in":"query","name":"page","required":false},{"schema":{"default":20,"type":"number"},"in":"query","name":"limit","required":false},{"schema":{"type":"string"},"in":"query","name":"search","required":false}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"users":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"job":{"anyOf":[{"type":"string"},{"type":"null"}]},"serviceId":{"type":"string"},"serviceName":{"anyOf":[{"type":"string"},{"type":"null"}]},"serviceDepartment":{"anyOf":[{"type":"string"},{"type":"null"}]},"role":{"anyOf":[{"type":"string"},{"type":"null"}]},"createdAt":{"type":"string"}},"required":["id","name","email","job","serviceId","serviceName","serviceDepartment","role","createdAt"]}},"total":{"type":"number"},"page":{"type":"number"},"limit":{"type":"number"}},"required":["users","total","page","limit"]}}}}}}},"/api/stats/public":{"get":{"parameters":[{"schema":{"type":"string"},"in":"query","name":"from","required":false},{"schema":{"type":"string"},"in":"query","name":"to","required":false}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"totalConstats":{"type":"number"},"totalReports":{"type":"number"},"sentConstats":{"type":"number"},"sentReports":{"type":"number"},"totalUsers":{"type":"number"},"usersWithNoDocuments":{"type":"number"},"activeUsersInPeriod":{"type":"number"},"deployedUdapCount":{"type":"number"},"deployedCrmhCount":{"type":"number"},"periodFrom":{"type":"string"},"periodTo":{"type":"string"}},"required":["totalConstats","totalReports","sentConstats","sentReports","totalUsers","usersWithNoDocuments","activeUsersInPeriod","deployedUdapCount","deployedCrmhCount","periodFrom","periodTo"]}}}}}}},"/api/stats/accounts":{"get":{"parameters":[{"schema":{"type":"string"},"in":"query","name":"from","required":false},{"schema":{"type":"string"},"in":"query","name":"to","required":false}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"month":{"type":"string"},"year":{"type":"number"},"count":{"type":"number"}},"required":["month","year","count"]}}}}}}}},"/api/stats/documents":{"get":{"parameters":[{"schema":{"type":"string"},"in":"query","name":"from","required":false},{"schema":{"type":"string"},"in":"query","name":"to","required":false}],"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"month":{"type":"string"},"year":{"type":"number"},"totalReports":{"type":"number"},"totalStateReports":{"type":"number"},"sentReports":{"type":"number"},"sentStateReports":{"type":"number"}},"required":["month","year","totalReports","totalStateReports","sentReports","sentStateReports"]}}}}}}}},"/api/stats/admin":{"get":{"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"object","properties":{"constatsByService":{"type":"array","items":{"type":"object","properties":{"serviceId":{"type":"string"},"serviceName":{"anyOf":[{"type":"string"},{"type":"null"}]},"sentConstats":{"type":"number"},"totalConstats":{"type":"number"}},"required":["serviceId","serviceName","sentConstats","totalConstats"]}},"abandonedConstats":{"type":"number"},"abandonedReports":{"type":"number"},"totalConstats":{"type":"number"},"totalReports":{"type":"number"},"totalUsers":{"type":"number"}},"required":["constatsByService","abandonedConstats","abandonedReports","totalConstats","totalReports","totalUsers"]}}}}}}},"/api/stats/udap":{"get":{"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"department":{"type":"string"},"totalStateReports":{"type":"number"},"sentStateReports":{"type":"number"},"totalReports":{"type":"number"},"sentReports":{"type":"number"},"usersCount":{"type":"number"}},"required":["department","totalStateReports","sentStateReports","totalReports","sentReports","usersCount"]}}}}}}}},"/api/stats/crmh":{"get":{"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"region":{"type":"string"},"totalStateReports":{"type":"number"},"sentStateReports":{"type":"number"},"totalReports":{"type":"number"},"sentReports":{"type":"number"},"usersCount":{"type":"number"}},"required":["region","totalStateReports","sentStateReports","totalReports","sentReports","usersCount"]}}}}}}}},"/api/stats/jobs":{"get":{"responses":{"200":{"description":"Default Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"job":{"anyOf":[{"type":"string"},{"type":"null"}]},"count":{"type":"number"}},"required":["job","count"]}}}}}}}},"/attachment/{id}/{filename}?":{"get":{"parameters":[{"schema":{"type":"string"},"in":"path","name":"id","required":true},{"schema":{"type":"string"},"in":"path","name":"filename","required":true}],"responses":{"200":{"description":"Default Response"}}}}}}